add: interface XMLParserBuilder
